+++ ext/pdo/pdo_stmt.c
@@ -2344,6 +2344,7 @@ static void row_dim_write(zend_object *object, zval *member, zval *value)
 	}
 }
 
+// todo: make row_prop_exists return bool as well
 static int row_prop_exists(zend_object *object, zend_string *name, int check_empty, void **cache_slot)
 {
 	pdo_row_t *row = (pdo_row_t *)object;
@@ -2366,8 +2367,7 @@ static int row_prop_exists(zend_object *object, zend_string *name, int check_emp
 	bool res = check_empty ? i_zend_is_true(retval) : Z_TYPE(tmp_val) != IS_NULL;
 	zval_ptr_dtor_nogc(retval);
 
-	// todo: make row_prop_exists return bool as well
-	return (int)res;
+	return res;
 }
 
 
@@ -2392,7 +2392,7 @@ static int row_dim_exists(zend_object *object, zval *offset, int check_empty)
 		ZEND_ASSERT(retval == &tmp_val);
 		bool res = check_empty ? i_zend_is_true(retval) : Z_TYPE(tmp_val) != IS_NULL;
 		zval_ptr_dtor_nogc(retval);
-		return (int)res;
+		return res;
 	} else {
 		zend_string *member = zval_try_get_string(offset);
 		if (!member) {
